+++ NEWS.md
@@ -6,7 +6,7 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.15.0 - 2026-01-13]
 
 ### Added
 
@@ -16,8 +16,6 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   Users now have the choice between multiple models. Note that
   depending on the model, the number of daily forecasts may vary; Also
   alerts and moon data aren't supported by Open-Meteo.
-
-### Changed
 
 ### Removed